% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panev.biomartSpecies.R
\name{panev.biomartSpecies}
\alias{panev.biomartSpecies}
\title{Gene list preparation species codes}
\usage{
panev.biomartSpecies(string = NULL)
}
\arguments{
\item{string}{A string used to search the organism code within the full list. If \emph{NULL}, a list with all available organisms is returned (default = NULL).}
}
\value{
A \emph{dataframe} with the list of organisms available for gene list preparation. The \emph{dataframe} contains three columns with \emph{organism_code}, \emph{description} and \emph{version} respectively.
}
\description{
The function retrieves the correct \emph{organism code} from \emph{ensembl.org} website. This code is mandatory to run \code{\link[PANEV]{panev.dataPreparation}} and \code{\link[PANEV]{panev.exprdataPreparation}}, useful to generate a properly formatted gene list for PANEV, switching up-to-date gene identifiers from \emph{ensembl} to \emph{entrez} annotation (or vice versa).
}
\details{
This function is based on \code{\link[biomaRt]{listDatasets}} function from \pkg{biomaRt} package (\url{http://bioconductor.org/packages/release/bioc/html/biomaRt.html}).
}
\examples{
##### EXAMPLES CODE #####
#Create a list of all available organisms
list <- panev.biomartSpecies(string = NULL)

#Look for a specific organism matching a search string 
list <- panev.biomartSpecies(string = "cow")
}
\references{
Mapping identifiers for the integration of genomic datasets with the R/Bioconductor package biomaRt. Steffen Durinck, Paul T. Spellman, Ewan Birney and Wolfgang Huber, Nature Protocols 4, 1184-1191 (2009).

BioMart and Bioconductor: a powerful link between biological databases and microarray data analysis. Steffen Durinck, Yves Moreau, Arek Kasprzyk, Sean Davis, Bart De Moor, Alvis Brazma and Wolfgang Huber, Bioinformatics 21, 3439-3440 (2005).
}
\author{
Valentino Palombo (\email{valentino.palombo@gmail.com})
}
