% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panev.dataPreparation.R
\name{panev.dataPreparation}
\alias{panev.dataPreparation}
\title{PANEV gene list preparation}
\usage{
panev.dataPreparation(in.file, gene_id = NULL, biomart.species = NULL)
}
\arguments{
\item{in.file}{Name of input file (with extension). The input file is a one column list of genes of interest, in \emph{ensembl} or \emph{entrez} gene annotation. The file \bold{must} rely in the working directory.}

\item{gene_id}{Type of gene identifiers provided. User can choose among \emph{ensembl} or \emph{entrez gene ID} annotation (default = NULL).}

\item{biomart.species}{The biomaRt organism code. The handy function \code{\link[PANEV]{panev.biomartSpecies}} is provided to retrieve the correct code (default = NULL).}
}
\value{
A \emph{<in.file>_converted.txt} file containing three columns with \emph{'ensembl_gene_id'}, \emph{'entrezgene'} and \emph{'external_gene_name'} respectively, stored in the working directory.
}
\description{
The function helps to create a properly formatted gene list required by \code{\link[PANEV]{panev.script}}: a \emph{dataframe} with three columns labelled as '\emph{ensembl_gene_id}', '\emph{entrezgene}' and '\emph{external_gene_name}', respectively.
}
\details{
This function is based on the main \pkg{biomaRt} (\url{http://bioconductor.org/packages/release/bioc/html/biomaRt.html}) query function \code{\link[biomaRt]{getBM}} which, given a set of filters and corresponding values, it retrieves the user specified attributes from the BioMart database one is connected to.
}
\examples{
##### EXAMPLES CODE #####
#Copy the example files in the current working directory
panev.example()

#Look for the organism code matching the search string 
list <- panev.biomartSpecies(string = "cow")
biomart.species <- as.character(list[1,1]) # btaurus_gene_ensembl

#Prepare PANEV input file using a gene list containing ensembl gene id.
genelist_converted <- panev.dataPreparation(in.file = "ensembl_genelist.txt", 
                                          gene_id = "ensembl", 
                                          biomart.species = biomart.species)

#Prepare PANEV input file using a gene list containing entraz gene id.
genelist_converted <- panev.dataPreparation(in.file = "entrez_genelist.txt", 
                                          gene_id = "entrez", 
                                          biomart.species = biomart.species)

}
\references{
Mapping identifiers for the integration of genomic datasets with the R/Bioconductor package biomaRt. Steffen Durinck, Paul T. Spellman, Ewan Birney and Wolfgang Huber, Nature Protocols 4, 1184-1191 (2009).

BioMart and Bioconductor: a powerful link between biological databases and microarray data analysis. Steffen Durinck, Yves Moreau, Arek Kasprzyk, Sean Davis, Bart De Moor, Alvis Brazma and Wolfgang Huber, Bioinformatics 21, 3439-3440 (2005).
}
\author{
Valentino Palombo (\email{valentino.palombo@gmail.com})
}
