% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panev.exprdataPreparation.R
\name{panev.exprdataPreparation}
\alias{panev.exprdataPreparation}
\title{PANEV gene expression dataset preparation}
\usage{
panev.exprdataPreparation(in.file, gene_id = NULL, biomart.species = NULL)
}
\arguments{
\item{in.file}{Name of input file (with extension). The input file is a table with three columns labelled \emph{ensembl_gene_id} OR \emph{'entrezgene'}, \emph{'FC'} and \emph{'pvalue'}, respectivelly. The input file \bold{must} rely in the working directory.}

\item{gene_id}{Type of annotation of provided gene list. You can choose among \emph{ensembl} or \emph{entrez} gene ID annotation (default = NULL).}

\item{biomart.species}{The biomaRt species code of interest. You can check the list of all available species on biomaRt with the function \code{\link[PANEV]{panev.biomartSpecies}} (default = NULL).}
}
\value{
A \emph{<in.file>_converted.txt} file with five columns labeled \emph{'ensembl_gene_id'}, \emph{'entrezgene'} and \emph{'external_gene_name'}, \emph{'FC'}, \emph{'pvalue'}, stored in the working directory.
}
\description{
In case of an expression dataset, to run \code{\link[PANEV]{panev.exprnetwork}} function, PANEV requires a \emph{dataframe} containing five columns labelled as follows: \emph{ensembl gene id}, \emph{entrezgene}, \emph{external gene name}, \emph{FC}, \emph{pvalue}. 
This function helps to create a properly formatted dataframe, converting the gene list from \emph{ensembl} or \emph{entrez} gene annotation (or vice versa) and adding the \emph{gene symbol}.
}
\details{
This function is based on the main \pkg{biomaRt} (\url{http://bioconductor.org/packages/release/bioc/html/biomaRt.html} query function \code{\link[biomaRt]{getBM}} which, given a set of filters and corresponding values, it retrieves the user specified attributes from the BioMart database one is connected to.
}
\examples{
##### EXAMPLES CODE #####
#Copy the example files in the current working directory
panev.example()

#Look for the organism code matching the search string 
list <- panev.biomartSpecies(string = "pig")
biomart.species <- as.character(list[6,1]) # sscrofa_gene_ensembl

#Convert the gene list with ensembl gene id
# Copy the example data file 'gene_list.txt' in the current working directory
genelist_converted <- panev.exprdataPreparation(in.file = "ensembl_expr_genelist.txt", 
                                              gene_id="ensembl", 
                                              biomart.species = biomart.species)

#Convert the gene list with entrez gene id
# Copy the example data file 'gene_list.txt' in the current working directory
genelist_converted <- panev.exprdataPreparation(in.file = "entrez_expr_genelist.txt", 
                                              gene_id="entrez", 
                                              biomart.species = biomart.species)

}
\references{
Mapping identifiers for the integration of genomic datasets with the R/Bioconductor package biomaRt. Steffen Durinck, Paul T. Spellman, Ewan Birney and Wolfgang Huber, Nature Protocols 4, 1184-1191 (2009).

BioMart and Bioconductor: a powerful link between biological databases and microarray data analysis. Steffen Durinck, Yves Moreau, Arek Kasprzyk, Sean Davis, Bart De Moor, Alvis Brazma and Wolfgang Huber, Bioinformatics 21, 3439-3440 (2005).
}
\author{
Valentino Palombo (\email{valentino.palombo@gmail.com})
}
