% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panev.exprnetwork.R
\name{panev.exprnetwork}
\alias{panev.exprnetwork}
\title{PANEV on gene expression datasets}
\usage{
panev.exprnetwork(in.file = NULL, path.file = NULL, out.file = "PANEV_expr", 
                      species = NULL, pvalue = 0.05)
}
\arguments{
\item{in.file}{Name of input file (with extension). The input file containing the differentially expressed genes (DEG) analysis results. The file \bold{must} contain five columns labeled \emph{'ensembl_gene_id'}, \emph{'entrezgene'}, \emph{'external_gene_name'}, \emph{'FC'} and \emph{'pvalue'}. The \code{\link[PANEV]{panev.exprdataPreparation}} function can be used to create the properly formatted input file from a gene expression list. The input file \bold{must} rely in working directory (default = NULL).}

\item{path.file}{Name of input file (with extension). The input file containing two columns labeled \emph{path_ID} and \emph{value}, containing respectively the ID of pathways of interest (\code{\link[PANEV]{panev.pathList}} function could be used to check if they are available), and relative biological estimated score value (e.g. flux value obtained with Dinamic Impact Approach analysis developed and described by Bionaz et al., 2012). The input file \bold{must} rely in working directory (default = NULL).}

\item{out.file}{Name of output diagram file and folder where results will be stored (default = "PANEV_expr").}

\item{species}{The KEGG organism code. The correct code for the species of interest can be retrieved with the handy \code{\link[PANEV]{panev.speciesCode}} function (default = NULL).}

\item{pvalue}{The p-value cut-off for Differentially Expressed (DE) analysis (default = \emph{0.05}).}
}
\value{
A \emph{<out.file>.html} file with the diagram visualization of PANEV results is created and stored in a folder named as \emph{PANEV_exprRESULTS_<out.name>}, created in the work directory.
}
\description{
The function visualizes PANEV results from expression gene datasets, considering any interactions among selected pathways. For each pathway the diagram helps to visualize associated genes that overcome a user defined \emph{p-value} threshold. The color is based on the fold change (\emph{FC}) value.
}
\details{
This function is based on the main KEGGREST query function \code{\link[KEGGREST]{keggLink}}.
}
\examples{
##### EXAMPLES CODE #####
#Copy the example files in the current working directory
panev.example()

#Parameters
in.file = "exprdata.txt"
path.file  = "expr_listPath.txt"
out.file = "expression_data"
pvalue = 0.05

#Look for the specie code matching the search string 
list <- panev.speciesCode(string = "scrofa")
species = as.character(list[1,2]) # ssc

# Run the PANEV function
panev.exprnetwork(in.file = in.file, 
                  path.file = path.file, 
                  out.file = out.file, 
                  species = species, 
                  pvalue = pvalue)

}
\references{
Bionaz, M., K. Periasamy, S.L. Rodriguez-Zas, W.L. Hurley, and J.J. Loor. 2012. A novel dynamic impact approach (DIA) for functional analysis of time-course omics studies: validation using the bovine mammary transcriptome. PLos One 7:e32455. doi:10.1371/journal.pone.0032455.

Tenenbaum D (2017). KEGGREST: Client-side REST access to KEGG. R package version 1.16.1.

Thieurmel B (2016). visNetwork: Network Visualization using 'vis.js' Library. R package version 2.0.3. https://CRAN.R-project.org/package=visNetwork
}
\author{
Valentino Palombo (\email{valentino.palombo@gmail.com})
}
