% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panev.network.enrichment.R
\name{panev.network.enrichment}
\alias{panev.network.enrichment}
\title{Enrichment analysis considering PANEV result as a background}
\usage{
panev.network.enrichment(in.file, out.file = "PANEV_enrich", species = NULL, FL = NULL, levels = 2)
}
\arguments{
\item{in.file}{Name of input file (with extension) containing the gene list of interest. The file \bold{must} contain three columns labelled as \emph{'ensembl_gene_id'}, \emph{'entrezgene'} and \emph{'external_gene_name'}, respectively. The file \bold{must} rely in the working directory. The handy function \code{\link[PANEV]{panev.dataPreparation}} could be used to create a properly formatted input file from a single gene list.}

\item{out.file}{Name of the folder where the results will be stored and of the output diagram file (default =  'PANEV_enrich').}

\item{species}{The code of your species of interest. The correct code can get among the list of those available in KEGG with the handy function \code{\link[PANEV]{panev.speciesCode}}.}

\item{FL}{A list of pathways of first level to investigate. The list of all available pathways can get with the \code{\link[PANEV]{panev.pathList}} function.}

\item{levels}{The number of levels of interactions (from \emph{1} to \emph{n}) investingated (default = 2).}
}
\value{
A \emph{<out.file>_enrichment.txt} file containing the results of functional enrichment analysis, based on a one-sided Fisher's exact test (hypergeometric test) and considering the PANEV network as a background.
}
\description{
The function perform an enrichment analysis considering as backgroud the union of pathways investigated by PANEV for generate network result.The result help to interpret the PANEV output allowing to identify the pathways most strongly associated with the input list of genes.
}
\details{
This function is based on \code{\link[KEGGREST]{keggList}} and \code{\link[KEGGREST]{keggLink}} functions of \pkg{KEGGREST} package (\url{http://bioconductor.org/packages/release/bioc/html/KEGGREST.html}.

The enrichment analysis is based on \code{\link[bc3net]{enrichment}} function of \pkg{bc3net} package (\url{https://cran.r-project.org/web/packages/bc3net/index.html}.
}
\examples{
##### EXAMPLES CODE #####
#Copy the example files in the current working directory
panev.example()

#Look for the organism code matching the search string 
list <- panev.biomartSpecies(string = "cow")
biomart.species <- as.character(list[1,1]) # btaurus_gene_ensembl

#Prepare PANEV input file using a gene list containing ensembl gene id.
genelist_converted <- panev.dataPreparation(in.file = "ensembl_genelist.txt", 
                                          gene_id = "ensembl", 
                                          biomart.species = biomart.species)

#Look for the specie code matching the search string 
list <- panev.speciesCode(string = "bos")
species <- as.character(list[1,2]) # bta

#Overview on input data 
panev.network.enrichment(in.file = in.file, 
              out.file = out.file, 
              species = species, 
              FL = FL, 
              levels = 3)

}
\references{
Tenenbaum D (2017). KEGGREST: Client-side REST access to KEGG. R package version 1.16.1.

Simoes R de M, Emmert-Streib F (2012). Bagging statistical network inference from large-scale gene expression data. PLOS ONE; 7: e33624. doi:10.1371/journal.pone.0033624
}
\author{
Valentino Palombo (\email{valentino.palombo@gmail.com})
}
