% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panev.network.R
\name{panev.network}
\alias{panev.network}
\title{Pathways Network Visualization (PANEV)}
\usage{
panev.network(in.file = NULL, out.file = "PANEV_gene", species = NULL, 
                  FL = NULL, levels = 2)
}
\arguments{
\item{in.file}{Name of input file (with extension) containing the gene list of interest. The file \bold{must} contain three columns labelled as \emph{'ensembl_gene_id'}, \emph{'entrezgene'} and \emph{'external_gene_name'}, respectively. The file \bold{must} rely in the working directory. The handy function \code{\link[PANEV]{panev.dataPreparation}} could be used to create a properly formatted input file from a single gene list.}

\item{out.file}{Name of the folder where the results will be stored and of the output diagram file (default =  'PANEV_gene').}

\item{species}{The code of your species of interest. The correct code can get among the list of those available in KEGG with the handy function \code{\link[PANEV]{panev.speciesCode}}.}

\item{FL}{A list of pathways of first level to investigate. The list of all available pathways can get with the \code{\link[PANEV]{panev.pathList}} function.}

\item{levels}{The number of levels of interactions (from \emph{1} to \emph{n}) investingated (default = 2).}
}
\value{
\itemize{
\item{The script generates from \emph{1} to \emph{n} '\emph{txt}' files, (with \emph{n} equal to the levels of interaction required for the investigation). 
Each file is named \emph{(n)}Lgenes, based on \emph{n} level analysed 
and contains the genes falling inside the pathways of the specific degree of interaction investigated.
Each file contains five columns with \emph{ensembl gene}, \emph{entrez gene}, \emph{gene name}, \emph{path description} and \emph{path id}. 
The files are stored in a folder named as \emph{PANEV_RESULTS_<out.name>}, created in the work directory.}
\item{A \emph{<out.file>.html} file with the diagram visualization of PANEV results. The file is stored in the same folder of \emph{.txt} files}
}
}
\description{
The function performs a pathway analysis taking into account both upstream and downstream dependent network of functional related genes, from 2 to \emph{n} degrees of interaction. This function generates \emph{n} '\emph{.txt}' files, one for each level of interaction, containing the candidate genes and the related pathways highlighted. Along with the tabular format results, the function gets also the diagram visualization of PANEV results, saved in an interactive '\emph{.html}' file.
}
\details{
For each level required for the investigation, a single \emph{.txt} file will be generated, containing the genes and the related pathways for a specific level of interaction. The function get also the diagram visualization of results. This function is based on the main \pkg{KEGGREST} package (\url{https://bioconductor.org/packages/release/bioc/html/KEGGREST.html}).
}
\examples{
##### EXAMPLES CODE #####
#Copy the example files in the current working directory
panev.example()

#Parameters
in.file="data.txt"
out.file="example"
FL = c("path:map00061", "path:map00062", "path:map00071", "path:map00072")
levels = 2

#Look for the specie code matching the search string 
list <- panev.speciesCode(string = "bos")
species=as.character(list[1,2]) # bta 

# Run the PANEV function
panev.network(in.file = in.file, 
              out.file = out.file, 
              species = species, 
              FL = FL, 
              levels = levels)
           
}
\references{
Tenenbaum D (2017). KEGGREST: Client-side REST access to KEGG. R package version 1.16.1.

Thieurmel B (2016). visNetwork: Network Visualization using 'vis.js' Library. R package version 2.0.3.  https://CRAN.R-project.org/package=visNetwork
}
\author{
Valentino Palombo (\email{valentino.palombo@gmail.com})
}
