% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panev.speciesCode.R
\name{panev.speciesCode}
\alias{panev.speciesCode}
\title{PANEV species codes}
\usage{
panev.speciesCode(string = NULL)
}
\arguments{
\item{string}{A string used to search the organism code within the full list. (default = NULL).}
}
\value{
A \emph{dataframe} containing two columns named \emph{'species'} and \emph{'panev code'}, respectively.
}
\description{
Get the list of species available for panev. The correct \emph{'panev code'} is mandatory for run correctly the main PANEV functions.
}
\details{
This function is based on the \code{\link[KEGGREST]{keggList}} function of \pkg{KEGGREST} package (\url{http://bioconductor.org/packages/release/bioc/html/KEGGREST.html}.
}
\examples{
##### EXAMPLES CODE #####
#Create a list of all available species for PANEV
list <- panev.speciesCode()

#Look for a specific species in PANEV, matching a search string
list <- panev.speciesCode(string="bos")
}
\references{
Tenenbaum D (2017). KEGGREST: Client-side REST access to KEGG. R package version 1.16.1.
}
\author{
Valentino Palombo (\email{valentino.palombo@gmail.com})
}
