% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panev.stats.enrichment.R
\name{panev.stats.enrichment}
\alias{panev.stats.enrichment}
\title{Data overview and enrichment analysis}
\usage{
panev.stats.enrichment(in.file, out.file = "PANEV_enrich", species = NULL)
}
\arguments{
\item{in.file}{Name of input file (with extension). The file \bold{must} contain three columns labelled \emph{'ensembl_gene_id'}, \emph{'entrezgene'} and \emph{'external_gene_name'} respectively. The file \bold{must} rely in the working directory. The handy function \code{\link[PANEV]{panev.dataPreparation}} could be used to create a properly formatted input file from a gene list.}

\item{out.file}{The specific name of generated files (default = 'PANEV_enrich').}

\item{species}{The species code of interest. You can get the correct code among the list of those available in KEGG with the \code{\link[PANEV]{panev.speciesCode}} function.}
}
\value{
A \emph{<out.file>_enrichment.txt} file containing the results of functional enrichment analysis, based on a one-sided Fisher's exact test (hypergeometric test).

A \emph{<out.file>_GxP.txt} file containing the numbers of gene(s) detected for single pathway.

A \emph{<out.file>_PxG.txt} file containing the numbers of pathway(s) detected for single gene.
}
\description{
The function takes as input the principal descriptive information on given gene list. In particular: number of gene(s) per single pathway, number of pathway(s) per single gene and enrichment analysis. When biological assumptions are broad or not strictly defined, pathways with more occurrences may be principal candidate pathways (FL) to be investigated with PANEV. The results obtained are based on KEGG Pathway database information.
}
\details{
This function is based on \code{\link[KEGGREST]{keggList}} and \code{\link[KEGGREST]{keggLink}} functions of \pkg{KEGGREST} package (\url{http://bioconductor.org/packages/release/bioc/html/KEGGREST.html}.
}
\examples{
##### EXAMPLES CODE #####
#Copy the example files in the current working directory
panev.example()

#Look for the organism code matching the search string 
list <- panev.biomartSpecies(string = "cow")
biomart.species <- as.character(list[1,1]) # btaurus_gene_ensembl

#Prepare PANEV input file using a gene list containing ensembl gene id.
genelist_converted <- panev.dataPreparation(in.file = "ensembl_genelist.txt", 
                                          gene_id = "ensembl", 
                                          biomart.species = biomart.species)

#Look for the specie code matching the search string 
list <- panev.speciesCode(string = "bos")
species <- as.character(list[1,2]) # bta

#Overview on input data 
panev.stats.enrichment(in.file = "ensembl_genelist_converted.txt", 
                     out.file = "example", 
                     species="bta")

}
\references{
Tenenbaum D (2017). KEGGREST: Client-side REST access to KEGG. R package version 1.16.1.
}
\author{
Valentino Palombo (\email{valentino.palombo@gmail.com})
}
